Refacor, move Kasner to equations
