--- nrpy/infrastructures/BHaH/kasner_solution/__init__.py
@@ -1,1 +1,0 @@
-from . import diagnostics, initial_data, kasner_exact
